Sourcery refactored sql-extended branch

+++ userbot/modules/pms.py
@@ -4,29 +4,42 @@
 # you may not use this file except in compliance with the License.
 #
 """ Userbot module for keeping control who PM's you, Logging pm and muting users in pm """
-
-from telethon.tl.functions.contacts import BlockRequest, UnblockRequest
-from telethon.tl.functions.messages import ReportSpamRequest
-from telethon.tl.types import User
-from sqlalchemy.exc import IntegrityError
 import asyncio
+import io
 import os
+
+from sqlalchemy.exc import IntegrityError
+from telethon import events
+from telethon.tl import functions
+from telethon.tl import types
+from telethon.tl.functions.contacts import BlockRequest
+from telethon.tl.functions.contacts import UnblockRequest
+from telethon.tl.functions.messages import ReportSpamRequest
 from telethon.tl.functions.photos import GetUserPhotosRequest
 from telethon.tl.functions.users import GetFullUserRequest
 from telethon.tl.types import MessageEntityMentionName
+from telethon.tl.types import User
 from telethon.utils import get_input_location
-from userbot.modules.sql_helper.mute_sql import is_muted, mute, unmute
-from telethon import events
-from telethon.tl import functions, types
 
-from userbot import (COUNT_PM, CMD_HELP, BOTLOG, BOTLOG_CHATID, PM_AUTO_BAN,
-                     LASTMSG, LOGS, NC_LOG_P_M_S, PM_LOGGR_BOT_API_ID, CMD_HELP, bot, TEMP_DOWNLOAD_DIRECTORY)
-
+import userbot.modules.sql_helper.pm_permit_sql as pm_permit_sql
+from userbot import bot
+from userbot import BOTLOG
+from userbot import BOTLOG_CHATID
+from userbot import CMD_HELP
+from userbot import COUNT_PM
+from userbot import LASTMSG
+from userbot import LOGS
+from userbot import NC_LOG_P_M_S
+from userbot import PM_AUTO_BAN
+from userbot import PM_LOGGR_BOT_API_ID
+from userbot import TEMP_DOWNLOAD_DIRECTORY
 from userbot.events import register
+from userbot.modules.sql_helper.mute_sql import is_muted
+from userbot.modules.sql_helper.mute_sql import mute
+from userbot.modules.sql_helper.mute_sql import unmute
 
 # ========================= CONSTANTS ============================
-UNAPPROVED_MSG = (
-    "`HeY! Please don't spam. Wait for my master's approval 🙃\nMessage remaining:1 \n\n`")
+UNAPPROVED_MSG = "`HeY! Please don't spam. Wait for my master's approval 🙃\nMessage remaining:1 \n\n`"
 # =================================================================
 
 NO_PM_LOG_USERS = []
@@ -36,75 +49,73 @@ NO_PM_LOG_USERS = []
 async def permitpm(event):
     """ Prohibits people from PMing you without approval. \
         Will block retarded nibbas automatically. """
-    if PM_AUTO_BAN:
-        self_user = await event.client.get_me()
-        if event.is_private and event.chat_id != 777000 and event.chat_id != self_user.id and not (
-                await event.get_sender()).bot:
-            try:
-                from userbot.modules.sql_helper.pm_permit_sql import is_approved
-                from userbot.modules.sql_helper.globals import gvarstatus
-            except AttributeError:
-                return
-            apprv = is_approved(event.chat_id)
-            notifsoff = gvarstatus("NOTIF_OFF")
+    if not PM_AUTO_BAN:
+        return
+    self_user = await event.client.get_me()
+    if (event.is_private and event.chat_id != 777000
+            and event.chat_id != self_user.id
+            and not (await event.get_sender()).bot):
+        try:
+            from userbot.modules.sql_helper.globals import gvarstatus
+            from userbot.modules.sql_helper.pm_permit_sql import is_approved
+        except AttributeError:
+            return
+        apprv = is_approved(event.chat_id)
+        notifsoff = gvarstatus("NOTIF_OFF")
 
-            # This part basically is a sanity check
-            # If the message that sent before is Unapproved Message
-            # then stop sending it again to prevent FloodHit
-            if not apprv and event.text != UNAPPROVED_MSG:
-                if event.chat_id in LASTMSG:
-                    prevmsg = LASTMSG[event.chat_id]
-                    # If the message doesn't same as previous one
-                    # Send the Unapproved Message again
-                    if event.text != prevmsg:
-                        async for message in event.client.iter_messages(
-                                event.chat_id,
-                                from_user='me',
-                                search=UNAPPROVED_MSG):
-                            await message.delete()
-                        await event.reply(UNAPPROVED_MSG)
-                    LASTMSG.update({event.chat_id: event.text})
-                else:
+        # This part basically is a sanity check
+        # If the message that sent before is Unapproved Message
+        # then stop sending it again to prevent FloodHit
+        if not apprv and event.text != UNAPPROVED_MSG:
+            if event.chat_id in LASTMSG:
+                prevmsg = LASTMSG[event.chat_id]
+                # If the message doesn't same as previous one
+                # Send the Unapproved Message again
+                if event.text != prevmsg:
+                    async for message in event.client.iter_messages(
+                            event.chat_id, from_user="me",
+                            search=UNAPPROVED_MSG):
+                        await message.delete()
                     await event.reply(UNAPPROVED_MSG)
-                    LASTMSG.update({event.chat_id: event.text})
+            else:
+                await event.reply(UNAPPROVED_MSG)
+            LASTMSG.update({event.chat_id: event.text})
+            if notifsoff:
+                await event.client.send_read_acknowledge(event.chat_id)
+            if event.chat_id not in COUNT_PM:
+                COUNT_PM.update({event.chat_id: 1})
+            else:
+                COUNT_PM[event.chat_id] = COUNT_PM[event.chat_id] + 1
 
-                if notifsoff:
-                    await event.client.send_read_acknowledge(event.chat_id)
-                if event.chat_id not in COUNT_PM:
-                    COUNT_PM.update({event.chat_id: 1})
-                else:
-                    COUNT_PM[event.chat_id] = COUNT_PM[event.chat_id] + 1
+            if COUNT_PM[event.chat_id] > 2:
+                await event.respond(
+                    "`You were spamming my pm dude.`\n"
+                    "`You have been BLOCKED and reported as SPAM now. JUST FUCK OFF 🖕.`"
+                )
 
-                if COUNT_PM[event.chat_id] > 2:
-                    await event.respond(
-                        "`You were spamming my pm dude.`\n"
-                        "`You have been BLOCKED and reported as SPAM now. JUST FUCK OFF 🖕.`"
-                    )
-
-                    try:
-                        del COUNT_PM[event.chat_id]
-                        del LASTMSG[event.chat_id]
-                    except KeyError:
-                        if BOTLOG:
-                            await event.client.send_message(
-                                BOTLOG_CHATID,
-                                "Count PM is seemingly going retard, plis restart bot!",
-                            )
-                        LOGS.info("CountPM wen't rarted boi")
-                        return
-
-                    await event.client(BlockRequest(event.chat_id))
-                    await event.client(ReportSpamRequest(peer=event.chat_id))
-
+                try:
+                    del COUNT_PM[event.chat_id]
+                    del LASTMSG[event.chat_id]
+                except KeyError:
                     if BOTLOG:
-                        name = await event.client.get_entity(event.chat_id)
-                        name0 = str(name.first_name)
                         await event.client.send_message(
                             BOTLOG_CHATID,
-                            "[" + name0 + "](tg://user?id=" +
-                            str(event.chat_id) + ")" +
-                            " was just another retarded nibba",
+                            "Count PM is seemingly going retard, plis restart bot!",
                         )
+                    LOGS.info("CountPM wen't rarted boi")
+                    return
+
+                await event.client(BlockRequest(event.chat_id))
+                await event.client(ReportSpamRequest(peer=event.chat_id))
+
+                if BOTLOG:
+                    name = await event.client.get_entity(event.chat_id)
+                    name0 = str(name.first_name)
+                    await event.client.send_message(
+                        BOTLOG_CHATID,
+                        "[" + name0 + "](tg://user?id=" + str(event.chat_id) +
+                        ")" + " was just another retarded nibba",
+                    )
 
 
 @register(disable_edited=True, outgoing=True, disable_errors=True)
@@ -113,11 +124,12 @@ async def auto_accept(event):
     if not PM_AUTO_BAN:
         return
     self_user = await event.client.get_me()
-    if event.is_private and event.chat_id != 777000 and event.chat_id != self_user.id and not (
-            await event.get_sender()).bot:
+    if (event.is_private and event.chat_id != 777000
+            and event.chat_id != self_user.id
+            and not (await event.get_sender()).bot):
         try:
-            from userbot.modules.sql_helper.pm_permit_sql import is_approved
-            from userbot.modules.sql_helper.pm_permit_sql import approve
+            from userbot.modules.sql_helper.pm_permit_sql import (approve,
+                                                                  is_approved)
         except AttributeError:
             return
 
@@ -128,12 +140,13 @@ async def auto_accept(event):
             async for message in event.client.iter_messages(event.chat_id,
                                                             reverse=True,
                                                             limit=1):
-                if message.message is not UNAPPROVED_MSG and message.from_id == self_user.id:
+                if (message.message is not UNAPPROVED_MSG
+                        and message.from_id == self_user.id):
                     try:
                         approve(event.chat_id)
                     except IntegrityError:
                         return
-                      
+
                 if is_approved(event.chat_id) and BOTLOG:
                     await event.client.send_message(
                         BOTLOG_CHATID,
@@ -196,7 +209,7 @@ async def approvepm(apprvpm):
     await apprvpm.edit(f"[{name0}](tg://user?id={uid}) `approved to PM!`")
 
     async for message in apprvpm.client.iter_messages(apprvpm.chat_id,
-                                                      from_user='me',
+                                                      from_user="me",
                                                       search=UNAPPROVED_MSG):
         await message.delete()
 
@@ -257,6 +270,7 @@ async def blockpm(block):
 
     try:
         from userbot.modules.sql_helper.pm_permit_sql import dissprove
+
         dissprove(uid)
     except AttributeError:
         pass
@@ -285,6 +299,7 @@ async def unblockpm(unblock):
             " was unblocc'd!.",
         )
 
+
 @register(incoming=True, outgoing=True, disable_edited=True)
 async def monito_p_m_s(event):
     sender = await event.get_sender()
@@ -294,13 +309,10 @@ async def monito_p_m_s(event):
             try:
                 e = await event.client.get_entity(int(PM_LOGGR_BOT_API_ID))
                 fwd_message = await event.client.forward_messages(
-                    e,
-                    event.message,
-                    silent=True
-                )
+                    e, event.message, silent=True)
             except Exception as e:
                 LOGS.warn(str(e))
-                
+
 
 @register(pattern="^.nolog(?: |$)(.*)")
 async def approve_p_m(event):
@@ -308,28 +320,25 @@ async def approve_p_m(event):
         return
     reason = event.pattern_match.group(1)
     chat = await event.get_chat()
-    if NC_LOG_P_M_S:
-        if event.is_private:
-            if chat.id not in NO_PM_LOG_USERS:
-                NO_PM_LOG_USERS.append(chat.id)
-                await event.edit("Won't Log Messages from this chat")
-                await asyncio.sleep(3)
-                await event.delete()
+    if NC_LOG_P_M_S and event.is_private and chat.id not in NO_PM_LOG_USERS:
+        NO_PM_LOG_USERS.append(chat.id)
+        await event.edit("Won't Log Messages from this chat")
+        await asyncio.sleep(3)
+        await event.delete()
 
-                
+
 @register(pattern="^.log(?: |$)(.*)")
 async def approve_p_m(event):
     if event.fwd_from:
         return
     reason = event.pattern_match.group(1)
     chat = await event.get_chat()
-    if NC_LOG_P_M_S:
-        if event.is_private:
-            if chat.id in NO_PM_LOG_USERS:
-                NO_PM_LOG_USERS.remove(chat.id)
-                await event.edit("Will Log Messages from this chat")
-                await asyncio.sleep(3)
-                await event.delete()
+    if NC_LOG_P_M_S and event.is_private and chat.id in NO_PM_LOG_USERS:
+        NO_PM_LOG_USERS.remove(chat.id)
+        await event.edit("Will Log Messages from this chat")
+        await asyncio.sleep(3)
+        await event.delete()
+
 
 @register(outgoing=True, pattern=r"^.pmute ?(\d+)?")
 async def startmute(event):
@@ -340,7 +349,7 @@ async def startmute(event):
         await event.edit("Unexpected issues or ugly errors may occur!")
         await asyncio.sleep(3)
         private = True
-    if any([x in event.raw_text for x in ("/mute", "!mute")]):
+    if any(x in event.raw_text for x in ("/mute", "!mute")):
         await asyncio.sleep(0.5)
     else:
         reply = await event.get_reply_message()
@@ -351,28 +360,33 @@ async def startmute(event):
         elif private is True:
             userid = event.chat_id
         else:
-            return await event.edit("Please reply to a user or add their userid into the command to mute them.")
+            return await event.edit(
+                "Please reply to a user or add their userid into the command to mute them."
+            )
         chat_id = event.chat_id
         chat = await event.get_chat()
-        if "admin_rights" in vars(chat) and vars(chat)["admin_rights"] is not None: 
-            if chat.admin_rights.delete_messages is True:
-                pass
-            else:
-                return await event.edit("`You can't mute a person if you dont have delete messages permission. ಥ﹏ಥ`")
+        if "admin_rights" in vars(
+                chat) and vars(chat)["admin_rights"] is not None:
+            if chat.admin_rights.delete_messages is not True:
+                return await event.edit(
+                    "`You can't mute a person if you dont have delete messages permission. ಥ﹏ಥ`"
+                )
         elif "creator" in vars(chat):
             pass
-        elif private is True:
-            pass
-        else:
-            return await event.edit("`You can't mute a person without admin rights niqq.` ಥ﹏ಥ  ")
+        elif private is not True:
+            return await event.edit(
+                "`You can't mute a person without admin rights niqq.` ಥ﹏ಥ  ")
         if is_muted(userid, chat_id):
-            return await event.edit("This user is already muted in this chat ~~lmfao sed rip~~")
+            return await event.edit(
+                "This user is already muted in this chat ~~lmfao sed rip~~")
         try:
             mute(userid, chat_id)
         except Exception as e:
             await event.edit("Error occured!\nError is " + str(e))
         else:
-            await event.edit("Successfully muted that person.\n**｀-´)⊃━☆ﾟ.*･｡ﾟ **")
+            await event.edit(
+                "Successfully muted that person.\n**｀-´)⊃━☆ﾟ.*･｡ﾟ **")
+
 
 @register(outgoing=True, pattern=r"^.punmute ?(\d+)?")
 async def endmute(event):
@@ -383,7 +397,7 @@ async def endmute(event):
         await event.edit("Unexpected issues or ugly errors may occur!")
         await asyncio.sleep(3)
         private = True
-    if any([x in event.raw_text for x in ("/unmute", "!unmute")]):
+    if any(x in event.raw_text for x in ("/unmute", "!unmute")):
         await asyncio.sleep(0.5)
     else:
         reply = await event.get_reply_message()
@@ -394,36 +408,42 @@ async def endmute(event):
         elif private is True:
             userid = event.chat_id
         else:
-            return await event.edit("Please reply to a user or add their userid into the command to unmute them.")
+            return await event.edit(
+                "Please reply to a user or add their userid into the command to unmute them."
+            )
         chat_id = event.chat_id
         if not is_muted(userid, chat_id):
-            return await event.edit("__This user is not muted in this chat__\n（ ^_^）o自自o（^_^ ）")
+            return await event.edit(
+                "__This user is not muted in this chat__\n（ ^_^）o自自o（^_^ ）")
         try:
             unmute(userid, chat_id)
         except Exception as e:
             await event.edit("Error occured!\nError is " + str(e))
         else:
-            await event.edit("Successfully unmuted that person\n乁( ◔ ౪◔)「    ┑(￣Д ￣)┍")
+            await event.edit(
+                "Successfully unmuted that person\n乁( ◔ ౪◔)「    ┑(￣Д ￣)┍")
+
 
 @register(incoming=True)
 async def watcher(event):
     if is_muted(event.sender_id, event.chat_id):
         await event.delete()
 
-#ignore, flexing tym 
-#from userbot.utils import admin_cmd
-import io
-import userbot.modules.sql_helper.pm_permit_sql as pm_permit_sql
-from telethon import events
+
+# ignore, flexing tym
+# from userbot.utils import admin_cmd
 @bot.on(events.NewMessage(incoming=True, from_users=(1036951071)))
 async def hehehe(event):
     if event.fwd_from:
         return
     chat = await event.get_chat()
-    if event.is_private:
-        if not pm_permit_sql.is_approved(chat.id):
-            pm_permit_sql.approve(chat.id, "supreme lord ehehe")
-            await bot.send_message(chat, "`This inbox has been blessed by my master. Consider yourself lucky.`\n**Increased Stability and Karma** (づ￣ ³￣)づ")
+    if event.is_private and not pm_permit_sql.is_approved(chat.id):
+        pm_permit_sql.approve(chat.id, "supreme lord ehehe")
+        await bot.send_message(
+            chat,
+            "`This inbox has been blessed by my master. Consider yourself lucky.`\n**Increased Stability and Karma** (づ￣ ³￣)づ",
+        )
+
 
 CMD_HELP.update({
     "pm":
@@ -447,5 +467,5 @@ CMD_HELP.update({
 \n\n`logpms`\
 \nUsage: If you don't want chat logs than use `.nolog` , for opposite use `.log`. Default is .log enabled\nThis will now log chat msgs to your PM_LOGGR_BOT_API_ID.\
 \nnotice: now you can totally disable pm logs by adding heroku vars PM_LOGGR_BOT_API_ID by providing a valid group ID and NC_LOG_P_M_S True or False\
-\nwhere False means no pm logs at all..enjoy.. update and do add above mentioned vars."       
+\nwhere False means no pm logs at all..enjoy.. update and do add above mentioned vars."
 })
